Add new settings.

+++ inc/admin/post-update.php
@@ -21,13 +21,18 @@ add_filter( 'post_updated_messages', 'scm_notice_after_update_post' );
  * @return void
  */
 function scm_update_post( $post_ID, $post_after, $post_before ) {
-	$post_url = get_permalink( $post_ID );
-	$cache_key = md5( parse_url( $post_url, PHP_URL_PATH ) );
 
-	$driver_type = get_option( 'scm_option_driver' );
-	$driver = scm_driver_factory( $driver_type );
+	$option_caching_status = get_option( 'scm_option_caching_status', 'disable' );
 
-	$driver->delete( $cache_key );
+	if ( 'enable' === $option_caching_status ) {
+		$post_url = get_permalink( $post_ID );
+		$cache_key = md5( parse_url( $post_url, PHP_URL_PATH ) );
+
+		$driver_type = get_option( 'scm_option_driver' );
+		$driver = scm_driver_factory( $driver_type );
+
+		$driver->delete( $cache_key );
+	}
 }
 
 /**
@@ -38,14 +43,18 @@ function scm_update_post( $post_ID, $post_after, $post_before ) {
  */
 function scm_notice_after_update_post( $messages ) {
 
-	$custom = '</div>';
-	$custom .= '<div class="notice notice-warning is-dismissible"><p>';
-	$custom .= '<strong>' . __( 'Cache Master', 'cache-master' ) . '</strong>: ' . __( 'Cache has been cleared.', 'cache-master' );
-	$custom .= '</p></div>';
-	$custom .= '<div>';
+	$option_caching_status = get_option( 'scm_option_caching_status', 'disable' );
 
-	$messages['post'][1] = $messages['post'][1] . $custom;
-	$messages['page'][1] = $messages['page'][1] . $custom;
+	if ( 'enable' === $option_caching_status ) {
+		$custom = '</div>';
+		$custom .= '<div class="notice notice-warning is-dismissible"><p>';
+		$custom .= '<strong>' . __( 'Cache Master', 'cache-master' ) . '</strong>: ' . __( 'Cache has been cleared.', 'cache-master' );
+		$custom .= '</p></div>';
+		$custom .= '<div>';
+	
+		$messages['post'][1] = $messages['post'][1] . $custom;
+		$messages['page'][1] = $messages['page'][1] . $custom;
+	}
 
 	return $messages;
 }
